fix: lazy db connection

+++ backend/src/controllers/petController.js
@@ -1,10 +1,17 @@
 import pets from "../data/pets.js";
 import { addAuditLog } from "../services/auditService.js";
 
-import { db } from "../config/db.js";
+import { getDB } from "../config/db.js";
 
 export const getAllPets = async (req, res) => {
   try {
+
+    const db = await getDB();
+
+    if (!db) {
+      return res.status(500).json({ message: "DB not connected" });
+    }
+
     const [rows] = await db.execute("SELECT * FROM pets");
     res.json(rows);
   } catch (err) {
@@ -18,6 +25,12 @@ export const createPet = async (req, res) => {
   try {
     const { name, species, breed, age } = req.body;
     const userEmail = req.user.sub;
+
+    const db = await getDB();
+
+    if (!db) {
+      return res.status(500).json({ message: "DB not connected" });
+    }
 
     await db.execute(
       "INSERT INTO pets (name, species, breed, age, status, createdBy) VALUES (?, ?, ?, ?, ?, ?)",
@@ -35,6 +48,12 @@ export const getMyPets = async (req, res) => {
   try {
     const userEmail = req.user.sub;
 
+    const db = await getDB();
+
+    if (!db) {
+      return res.status(500).json({ message: "DB not connected" });
+    }
+
     const [rows] = await db.execute(
       "SELECT * FROM pets WHERE createdBy = ?",
       [userEmail]
@@ -49,6 +68,12 @@ export const getMyPets = async (req, res) => {
 export const deletePet = async (req, res) => {
   try {
     const { id } = req.params;
+
+    const db = await getDB();
+
+    if (!db) {
+      return res.status(500).json({ message: "DB not connected" });
+    }
 
     await db.execute("DELETE FROM pets WHERE id = ?", [id]);
